+++ packages/coinstac-manager/LICENSE
