+++ packages/coinstac-manager/README.md
