+++ packages/coinstac-manager/package.json
